# Minor speedup for creation of constructor and method injector